# src/swagger/swagger.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Fortress API",
		"version": "1.0.0",
		"description": "Fortress is a social media platform that allows men to connect, share experiences, and support each other in a safe and empowering environment. The platform is designed to foster positive masculinity, mental health awareness, and community engagement through various features such as posts, comments, and user profiles. This is achieved through anonimity, robust moderation, and a focus on constructive interactions."
	},
	"paths": {
...
		"/comment/post/{postId}": {
			"post": {
				"tags": ["Comment"],
				"summary": "Create comment on post",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCommentDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Comment created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentResponse"
								}
							}
						}
					},
					"400": {
						"description": "Validation error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			},
			"get": {
				"tags": ["Comment"],
				"summary": "Get comments for a post",
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20
						}
					},
					{
						"name": "parentCommentId",
						"in": "query",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Comments list",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentsListResponse"
								}
							}
						}
					}
				}
			}
		},
...
		"/post/{id}": {
			"get": {
				"tags": ["Post"],
				"summary": "Get post by id",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "Post",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": ["Post"],
				"summary": "Update post",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePostDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Post updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Post"],
				"summary": "Delete post",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GenericMessageResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/saved-post": {
			"post": {
				"tags": ["SavedPost"],
				"summary": "Save a post",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SavePostDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Saved",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SavedPostResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			},
			"get": {
				"tags": ["SavedPost"],
				"summary": "Get saved posts for current user",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20
						}
					},
					{
						"name": "collectionName",
						"in": "query",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Saved posts list",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SavedPostsListResponse"
								}
							}
						}
					}
				}
			}
		},
		"/saved-post/{postId}": {
			"delete": {
				"tags": ["SavedPost"],
				"summary": "Unsave a post",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Unsaved",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GenericMessageResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			},
			"/saved-post/collections": {
				"get": {
					"tags": ["SavedPost"],
					"summary": "Get collections",
					"security": [
						{
							"bearerAuth": []
						}
					],
					"responses": {
						"200": {
							"description": "Collections",
							"content": {
								"application/json": {
									"schema": {
										"type": "array",
										"items": {
											"type": "object",
											"properties": {
												"name": {
													"type": "string"
												},
												"count": {
													"type": "integer"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"/saved-post/check/{postId}": {
					"get": {
						"tags": ["SavedPost"],
						"summary": "Check if post is saved",
						"security": [
							{
								"bearerAuth": []
							}
						],
						"parameters": [
							{
								"name": "postId",
								"in": "path",
								"required": true,
								"schema": {
									"type": "string"
								}
							}
						],
						"responses": {
							"200": {
								"description": "Check result",
								"content": {
									"application/json": {
										"schema": {
											"type": "object",
											"properties": {
												"saved": {
													"type": "boolean"
												}
											}
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/story": {
			"post": {
				"tags": ["Story"],
				"summary": "Create story",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStoryDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Story created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StoryResponse"
								}
							}
						}
					},
					"400": {
						"description": "Validation error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			},
			"get": {
				"tags": ["Story"],
				"summary": "Get stories (feed)",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20
						}
					},
					{
						"name": "userId",
						"in": "query",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Stories list",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Story"
									}
								}
							}
						}
					}
				}
			}
		},
		"/story/user/{userId}": {
			"get": {
				"tags": ["Story"],
				"summary": "Get user stories",
				"parameters": [
					{
						"name": "userId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "User stories",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Story"
									}
								}
							}
						}
					}
				}
			},
			"/story/{id}": {
				"get": {
					"tags": ["Story"],
					"summary": "Get story by id",
					"parameters": [
						{
							"name": "id",
							"in": "path",
							"required": true,
							"schema": {
								"type": "string"
							}
						}
					],
					"security": [
						{
							"bearerAuth": []
						}
					],
					"responses": {
						"200": {
							"description": "Story",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/StoryResponse"
									}
								}
							}
						},
						"404": {
							"description": "Not found",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/ErrorResponse"
									}
								}
							}
						},
						"410": {
							"description": "Expired",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/ErrorResponse"
									}
								}
							}
						}
					}
				},
				"post": {
					"tags": ["Story"],
					"summary": "View a story",
					"parameters": [
						{
							"name": "id",
							"in": "path",
							"required": true,
							"schema": {
								"type": "string"
							}
						}
					],
					"security": [
						{
							"bearerAuth": []
						}
					],
					"responses": {
						"200": {
							"description": "Viewed",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/StoryViewResponse"
									}
								}
							}
						},
						"404": {
							"description": "Not found",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/ErrorResponse"
									}
								}
							}
						},
						"401": {
							"description": "Unauthorized",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/ErrorResponse"
									}
								}
							}
						}
					}
				},
				"delete": {
					"tags": ["Story"],
					"summary": "Delete story",
					"parameters": [
						{
							"name": "id",
							"in": "path",
							"required": true,
							"schema": {
								"type": "string"
							}
						}
					],
					"security": [
						{
							"bearerAuth": []
						}
					],
					"responses": {
						"200": {
							"description": "Deleted",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/GenericMessageResponse"
									}
								}
							}
						},
						"403": {
							"description": "Forbidden",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/ErrorResponse"
									}
								}
							}
						},
						"401": {
							"description": "Unauthorized",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/ErrorResponse"
									}
								}
							}
						}
					}
				}
			},
			"/story/{id}/viewers": {
				"get": {
					"tags": ["Story"],
					"summary": "Get story viewers",
					"parameters": [
						{
							"name": "id",
							"in": "path",
							"required": true,
							"schema": {
								"type": "string"
							}
						}
					],
					"security": [
						{
							"bearerAuth": []
						}
					],
					"responses": {
						"200": {
							"description": "Viewers list",
							"content": {
								"application/json": {
									"schema": {
										"type": "object",
										"properties": {
											"viewers": {
												"type": "array",
												"items": {
													"$ref": "#/components/schemas/User"
												}
											},
											"totalViews": {
												"type": "integer"
											}
										}
									}
								}
							}
						},
						"403": {
							"description": "Forbidden",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/ErrorResponse"
									}
								}
							}
						},
						"401": {
							"description": "Unauthorized",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/ErrorResponse"
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/user/{id}": {
			"get": {
				"tags": ["User"],
				"summary": "Get user by id",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "User found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					},
					"403": {
						"description": "Private profile",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"User": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string",
						"format": "email"
					},
					"profile": {
						"$ref": "#/components/schemas/UserProfile"
					},
					"counts": {
						"$ref": "#/components/schemas/UserCounts"
					},
					"isVerified": {
						"type": "boolean"
					},
					"isPrivate": {
						"type": "boolean"
					},
					"lastLoginAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"UserProfile": {
				"type": "object",
				"properties": {
					"fullName": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					},
					"website": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"gender": {
						"type": "string",
						"enum": ["male", "female", "other", "prefer_not_to_say"]
					}
				}
			},
			"UserCounts": {
				"type": "object",
				"properties": {
					"posts": {
						"type": "integer"
					},
					"followers": {
						"type": "integer"
					},
					"following": {
						"type": "integer"
					}
				}
			},
			"ErrorResponse": {
				"type": "object",
				"properties": {
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"okay": {
						"type": "boolean"
					}
				}
			},
			"ValidationErrorResponse": {
				"type": "object",
				"properties": {
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"error": {
						"type": "string"
					},
					"okay": {
						"type": "boolean"
					}
				}
			},
			"CreateCommentDto": {
				"type": "object",
				"required": ["text"],
				"properties": {
					"text": {
						"type": "string"
					},
					"parentCommentId": {
						"type": "string"
					},
					"mentions": {
						"type": "array",
						"items": {
							"type": "string"
						}
					}
				}
			},
...
			"Comment": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string"
					},
					"postId": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"userAvatar": {
						"type": "string"
					},
					"text": {
						"type": "string"
					},
					"parentCommentId": {
						"type": "string"
					},
					"mentions": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"CommentsListResponse": {
				"type": "object",
				"properties": {
					"comments": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Comment"
						}
					},
					"pagination": {
						"type": "object",
						"properties": {
							"page": {
								"type": "integer"
							},
							"limit": {
								"type": "integer"
							},
							"total": {
								"type": "integer"
							},
							"totalPages": {
								"type": "integer"
							}
						}
					}
				}
			},
			"CommentResponse": {
				"type": "object",
				"properties": {
					"okay": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"$ref": "#/components/schemas/Comment"
					}
				}
			},
			"GenericMessageResponse": {
				"type": "object",
				"properties": {
					"okay": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"type": "string"
					}
				}
			},
...
			"Post": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string"
					},
					"caption": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"userAvatar": {
						"type": "string"
					},
					"media": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/MediaDto"
						}
					},
					"location": {
						"$ref": "#/components/schemas/LocationDto"
					},
					"hashtags": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"counts": {
						"$ref": "#/components/schemas/PostCounts"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
...
			"CoordinatesDto": {
				"type": "object",
				"properties": {
					"latitude": {
						"type": "number"
					},
					"longitude": {
						"type": "number"
					}
				}
			},
			"LocationDto": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"coordinates": {
						"$ref": "#/components/schemas/CoordinatesDto"
					}
				}
			},
			"MediaDto": {
				"type": "object",
				"properties": {
					"type": {
						"type": "string",
						"enum": ["image", "video"]
					},
					"url": {
						"type": "string"
					},
					"thumbnailUrl": {
						"type": "string"
					},
					"width": {
						"type": "number"
					},
					"height": {
						"type": "number"
					},
					"duration": {
						"type": "number"
					}
				}
			},
...
			"UpdatePostDto": {
				"type": "object",
				"properties": {
					"caption": {
						"type": "string"
					},
					"location": {
						"$ref": "#/components/schemas/LocationDto"
					},
					"hashtags": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"mentions": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"isCommentsEnabled": {
						"type": "boolean"
					},
					"isLikesVisible": {
						"type": "boolean"
					}
				}
			},
			"PostCounts": {
				"type": "object",
				"properties": {
					"likes": {
						"type": "integer"
					},
					"comments": {
						"type": "integer"
					}
				}
			},
			"PostResponse": {
				"type": "object",
				"properties": {
					"okay": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"$ref": "#/components/schemas/Post"
					}
				}
			},
			"SavePostDto": {
				"type": "object",
				"required": ["postId"],
				"properties": {
					"postId": {
						"type": "string"
					},
					"collectionName": {
						"type": "string"
					}
				}
			},
			"SavedPost": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"postId": {
						"type": "string"
					},
					"collectionName": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"SavedPostResponse": {
				"type": "object",
				"properties": {
					"okay": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"$ref": "#/components/schemas/SavedPost"
					}
				}
			},
			"SavedPostsListResponse": {
				"type": "object",
				"properties": {
					"savedPosts": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/SavedPost"
						}
					},
					"pagination": {
						"type": "object",
						"properties": {
							"page": {
								"type": "integer"
							},
							"limit": {
								"type": "integer"
							},
							"total": {
								"type": "integer"
							},
							"totalPages": {
								"type": "integer"
							}
						}
					}
				}
			},
...
			"UserResponse": {
				"type": "object",
				"properties": {
					"okay": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"$ref": "#/components/schemas/User"
					}
				}
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
...
	"security": [{ "bearerAuth": [] }]
}